adding new query to get all the blogs "server".
adding new component for home page like blogcard, trending, and adding hooks to get all the blogs and the user blogs. also graphql queries to get the data from the server and storing the data to the redux state. adding new reducer to handle and store all the blogs received from the server

adding styles to the home page